Backend/mcp/google_calender/requirements.txt
mcp
httpx
starlette
python-dotenv
sqlalchemy
psycopg2-binary
